+++ src/backend/src/entities/dispute.entity.ts
