+++ src/backend/src/entities/notification.entity.ts
