+++ src/backend/src/entities/statistic.entity.ts
