+++ src/backend/src/entities/user.entity.ts
